Move debug module to api_extensions
