+++ frontend/catalyst/__init__.py
@@ -82,7 +82,6 @@ sys.modules["mlir_quantum._mlir_libs._quantumDialects.mitigation"] = types.Modul
     "mlir_quantum._mlir_libs._quantumDialects.mitigation"
 )
 
-from catalyst import debug
 from catalyst.api_extensions import *
 from catalyst.api_extensions import __all__ as _api_extension_list
 from catalyst.autograph import autograph_source
@@ -177,7 +176,6 @@ while processing the following with AutoGraph:
 __all__ = (
     "qjit",
     "QJIT",
-    "debug",
     "autograph_source",
     "autograph_ignore_fallbacks",
     "autograph_strict_conversion",
