+++ frontend/catalyst/api_extensions/__init__.py
@@ -16,6 +16,7 @@
 This module is a collection of public API extensions for pragramming with Catalyst frontends.
 """
 
+from catalyst.api_extensions import debug
 from catalyst.api_extensions.callbacks import pure_callback
 from catalyst.api_extensions.control_flow import cond, for_loop, while_loop
 from catalyst.api_extensions.differentiation import (
@@ -30,6 +31,7 @@ from catalyst.api_extensions.function_maps import vmap
 from catalyst.api_extensions.quantum_operators import adjoint, ctrl, measure
 
 __all__ = (
+    "debug",
     "pure_callback",
     "cond",
     "for_loop",
