+++ frontend/test/pytest/test_debug.py
@@ -18,8 +18,12 @@ import pennylane as qml
 import pytest
 
 from catalyst import debug, for_loop, qjit
+from catalyst.api_extensions.debug import (
+    compile_from_mlir,
+    get_cmain,
+    print_compilation_stage,
+)
 from catalyst.compiler import CompileOptions, Compiler
-from catalyst.debug import compile_from_mlir, get_cmain, print_compilation_stage
 from catalyst.utils.exceptions import CompileError
 from catalyst.utils.runtime import get_lib_path
